move ui types to utils
