+++ client/components/table/index.tsx
@@ -1,3 +1,2 @@
 export { Table } from "./table";
 export { useNullableTableSort } from "./table-sort";
-export { SortDirection, SortRequestHandler, TableSort } from "./types";
