+++ client/components/user/index.ts
@@ -1,2 +1,1 @@
-export { UserInfo } from "./types";
 export { User } from "./user";
