--- client/components/user/types.ts
@@ -1,4 +1,0 @@
-export type UserInfo = {
-  avatar: string;
-  name: string;
-};
