+++ client/components/user/user.tsx
@@ -1,8 +1,8 @@
 import block from "bem-cn";
 import React, { FC, memo } from "react";
+import { UserInfo } from "../../utils/ui-types";
 import { Avatar } from "../avatar";
 import { Text } from "../text";
-import { UserInfo } from "./types";
 
 import "./user.css";
